# Halve the number of calls to cosd, sind in spherical `cellarea`